ビューを渡せるようにした
