+++ src/js/game-object/shot/lightning-shot/index.ts
@@ -1,34 +1,15 @@
-import * as THREE from "three";
-
 import { ResourcesContainer } from "../../../resource";
-import { createLightningShotProps } from "./props/create-lightning-shot-props";
-import { LightningShotProps } from "./props/lightning-shot-props";
+import { LightningShot } from "./lightning-shot";
+import { PlayerLightningShotView } from "./view/player-lightning-shot-view";
 
-/** 電撃ショット */
-export class LightningShot {
-  /** プロパティ */
-  #props: LightningShotProps;
-
-  /**
-   * コンストラクタ
-   * @param options オプション
-   */
-  constructor(options: ResourcesContainer) {
-    this.#props = createLightningShotProps(options);
-  }
-
-  /**
-   * デストラクタ相当の処理
-   */
-  destructor(): void {
-    this.#props.mesh.destructor();
-  }
-
-  /**
-   * シーンに追加するオブジェクトを取得する
-   * @returns シーンに追加するオブジェクト
-   */
-  getObject3D(): THREE.Object3D {
-    return this.#props.mesh.getObject3D();
-  }
+/**
+ * プレイヤーの電撃ショットを生成する
+ * @param options オプション
+ * @returns プレイヤーの電撃ショット
+ */
+export function playerLightningShot(
+  options: ResourcesContainer,
+): LightningShot {
+  const view = new PlayerLightningShotView(options);
+  return new LightningShot({ ...options, view });
 }
