+++ src/js/game-object/shot/lightning-shot/props/lightning-shot-props.ts
@@ -1,5 +1,6 @@
 import { HorizontalAnimationMesh } from "../../../../mesh/horizontal-animation";
 import { LightningShotModel } from "../model/lightning-shot-model";
+import { LightningShotView } from "../view/lightning-shot-view";
 
 /** 電撃ショットのプロパティ */
 export type LightningShotProps = {
@@ -7,4 +8,6 @@ export type LightningShotProps = {
   readonly mesh: HorizontalAnimationMesh;
   /** モデル */
   model: LightningShotModel;
+  /** ビュー */
+  readonly view: LightningShotView;
 };
